funcionando

+++ Libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: sonheres <sonheres@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 13:19:46 by sonheres          #+#    #+#             */
-/*   Updated: 2023/09/26 18:27:57 by sonheres         ###   ########.fr       */
+/*   Updated: 2023/09/26 18:36:54 by sonheres         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 void	*ft_memset(void *b, int c, size_t len)
 {
 	char	*n;
-	int	i;
+	size_t	i;
 
 	n = b;
 	i = 0;
@@ -25,12 +25,12 @@ void	*ft_memset(void *b, int c, size_t len)
 		n[i] = c;
 		i++;
 	}
-	return(b);
+	return (b);
 }
 
 int	main(void)
 {
-	char cadena[] = "Cambio a cero";
+	char	cadena[] = "Cambio a cero";
 	char cadena_ft[] = "Cambio a cero";
 	
 	memset(cadena, '0', 30);
